+++ README.md
@@ -161,6 +161,7 @@ Purpose: Demonstrate NeMo Gym patterns and concepts.
 | Name               | Demonstrates                         | Config                                                                                                                             | README                                                                      |
 | ------------------ | ------------------------------------ | ---------------------------------------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------------- |
 | Multi Step         | Multi-step tool calling              | <a href='resources_servers/example_multi_step/configs/example_multi_step.yaml'>example_multi_step.yaml</a>                         | <a href='resources_servers/example_multi_step/README.md'>README</a>         |
+| Multi Turn         | Knowledge example                    | <a href='resources_servers/example_multi_turn/configs/example_multi_turn.yaml'>example_multi_turn.yaml</a>                         | <a href='resources_servers/example_multi_turn/README.md'>README</a>         |
 | Session State Mgmt | Session state management (in-memory) | <a href='resources_servers/example_session_state_mgmt/configs/example_session_state_mgmt.yaml'>example_session_state_mgmt.yaml</a> | <a href='resources_servers/example_session_state_mgmt/README.md'>README</a> |
 | Single Tool Call   | Basic single-step tool calling       | <a href='resources_servers/example_single_tool_call/configs/example_single_tool_call.yaml'>example_single_tool_call.yaml</a>       | <a href='resources_servers/example_single_tool_call/README.md'>README</a>   |
 <!-- END_EXAMPLE_ONLY_SERVERS_TABLE -->
