+++ resources_servers/example_multi_turn/configs/example_multi_turn.yaml
@@ -1,6 +1,6 @@
-example_multi_turn_env:
+example_multi_turn:
   resources_servers:
-    example_multi_turn_env:
+    example_multi_turn:
       entrypoint: app.py
       domain: knowledge
       verified: false
@@ -10,7 +10,7 @@ example_multi_turn_gymnasium_agent:
       entrypoint: app.py
       env_server:
         type: resources_servers
-        name: example_multi_turn_env
+        name: example_multi_turn
       model_server:
         type: responses_api_models
         name: policy_model
@@ -18,4 +18,4 @@ example_multi_turn_gymnasium_agent:
       datasets:
       - name: example
         type: example
-        jsonl_fpath: resources_servers/example_multi_turn_env/data/example.jsonl
+        jsonl_fpath: resources_servers/example_multi_turn/data/example.jsonl
