+++ resources_servers/example_multi_turn/data/example_metrics.json
@@ -1,7 +1,7 @@
 {
     "name": "example",
     "type": "example",
-    "jsonl_fpath": "resources_servers/example_multi_turn_env/data/example.jsonl",
+    "jsonl_fpath": "resources_servers/example_multi_turn/data/example.jsonl",
     "num_repeats": 1,
     "gitlab_identifier": null,
     "huggingface_identifier": null,
